Added tabs to textarea component and made some quality of life improvements.
Was able to get the tabs to work with the textarea component using the useRef hook
from React. The other changes were the formatting of the props for the dropdown components.
Decided to make them objects and use the spread syntax to reduce redundency. I've also decided to
use the Array.map method for the dropdown menus and decided to replace them with tables where the rows
are were the options are going to be listed.