Added menu options and Zoom In/Out functionality.
Reverted to older design where I just hardcoded in the table rows. This is just
because I though I wouldn't be able to put the onClick events into the rows but I've
realized that it is possible with the syntax that I had previously. I will probably
change this to make it look much better.